split and rename net_starter
